feat(plugins/plugin-client-common): export the command line rendering component
This PR introduces a component to render bash pipelines

part of #7365
